DRILL-7261: Simplify Easy framework config for new scan
Most format plugins are created using the Easy format plugin. A recent
change added support for the "row set" scan framework. After converting
the text and log reader plugins, it became clear that the setup code
could be made simpler.

* Add the user name to the "file scan" framework.
* Pass the file system, split and user name to the batch reader via
  the "schema negotiator" rather than via the constructor.
* Create the traditional "scan batch" scan or the new row-set scan via
  functions instead of classes.
* Add Easy config option and method to choose the kind of scan
  framework.
* Add Easy config options for some newer options such as whether the
  plugin supports statistics.

Simplified reader creation

* The batch reader can be created just by overriding a method.
* A default error context is provided if the plugin does not provide
  one.

Tested by running all unit tests for the CSV reader which is based on
the new framework, and by testing the converted log reader (that reader
is not part of this commit.)

closes #1796