DRILL-7222: Visualize estimated and actual row counts for a query
With statistics in place, it is useful to have the estimated rowcount along side the actual rowcount query profile's operator overview. A toggle button allows this with the estimated rows hidden by default

We can extract this from the Physical Plan section of the profile.
Added a toggle-ready table-column header

closes #1779